Function doc, db schema and changes in mig.templ.
Added docstrings to some of the functions of the core ns.

Now we need to know about the schema name to generate foreign keys
correctly.

Some functions were renamed to express the intent better.

The function `column-data` was changed slightly to give information
about the column default value and if it can be null or not. Something
that we need to do in the future is to add support for values of
decimals and floats as we do for `enum` columns.

Overall, the code needs some tweaks, separate concerns between more
namespaces, introduce a decent command line interface implementation and
provide a DAG to create the migrations in dependency order.